Rename sprite file
