+++ src/__init__.py
@@ -6,5 +6,5 @@ from .event_handler import EventHandler
 from .game import Game
 from .scene_2d import Scene2D
 from .scene_manager import SceneManager
-from .sprite_sheet import SpriteSheet
+from .sprites import SpriteSheet
 from .window import WindowDisplay, WindowScreen
